新增编译结果输出

+++ app/Http/Controllers/RunController.php
@@ -24,24 +24,43 @@ class RunController extends Controller
         putenv('PATH=' . env('PATH'));
 
         $result = null;
+        $error = null;
         if ($language == 'php') {
             $result = shell_exec("php $tmpfname");
         } else if ($language == 'c语言') {
             $sourceFile = "$tmpfname.c";
             $outFile = "$tmpfname.out";
             $result = shell_exec("cp $tmpfname $sourceFile");
-            $result = shell_exec("gcc $sourceFile -o $outFile");
-            $result = shell_exec($outFile);
+            exec("gcc $sourceFile -o $outFile 2>&1", $output, $retval);
             unlink($sourceFile);
-            unlink($outFile);
+
+            if($retval !== 0) {
+                $error = join("\n", $output);
+                $files = explode('/', $sourceFile);
+                $error = $files[count($files) - 1] . explode($sourceFile, $error)[1];
+                $result = null;
+            } else {
+                $result = shell_exec($outFile);
+                unlink($outFile);
+            }
         } else if ($language == 'c++') {
             $sourceFile = "$tmpfname.cpp";
             $outFile = "$tmpfname.out";
             $result = shell_exec("cp $tmpfname $sourceFile");
             $result = shell_exec("g++ $sourceFile -o $outFile");
-            $result = shell_exec($outFile);
+
+            exec("g++ $sourceFile -o $outFile 2>&1", $output, $retval);
             unlink($sourceFile);
-            unlink($outFile);
+
+            if($retval !== 0) {
+                $error = join("\n", $output);
+                $files = explode('/', $sourceFile);
+                $error = $files[count($files) - 1] . explode($sourceFile, $error)[1];
+                $result = null;
+            } else {
+                $result = shell_exec($outFile);
+                unlink($outFile);
+            }
         } else if ($language == 'java') {
             $result = shell_exec("java $tmpfname");
         } else if ($language == 'python2.7') {
@@ -53,7 +72,8 @@ class RunController extends Controller
         unlink($tmpfname);
 
         return response()->json([
-            'result' => $result
+            'result' => htmlentities($result),
+            'error' => $error
         ]);
     }
 }
